Move common functions into `common.c`
